backend/requirements.txt
aioredis==1.2.0
aniso8601==3.0.2
asgiref==3.2.2
asn1crypto==0.24.0
async-timeout==3.0.1
atomicwrites==1.3.0
attrs==19.1.0
autobahn==19.9.2
Automat==0.7.0
Babel==2.7.0
beautifulsoup4==4.8.0
bleach==3.1.0
bootstrap3-datetime==2.4
cachetools==3.1.1
certifi==2019.6.16
cffi==1.12.3
channels==2.2.0
channels-redis==2.4.0
chardet==3.0.4
chatter==0.1.3
constantly==15.1.0
cryptography==2.7
daphne==2.3.0
dj-database-url==0.5.0
Django==2.0
django-analytical==2.5.0
django-appconf==1.0.3
django-babel==0.6.2
django-bootstrap-datepicker-plus==3.0.5
django-bootstrap-form==3.4
django-bootstrap3==11.1.0
django-bootstrap4==1.0.1
django-ckeditor==5.7.1
django-contrib-comments==1.9.1
django-countries==5.3.3
django-crispy-forms==1.7.2
django-debug-toolbar==1.11
django-environ==0.4.5
django-filter==2.1.0
django-filters==0.2.1
django-fluent-comments==2.1
#django-geoip==0.5.2.1
django-guardian==1.5.0
django-haystack==2.8.1
django-heroku==0.3.1
django-imagekit==4.0.2
django-ipware==2.1.0
django-js-asset==1.2.2
django-markdownx==2.0.28
#django-pagedown==0.1.1
django-phonenumber-field==2.4.0
django-tag-parser==3.1
django-taggit==1.1.0
django-taggit-serializer==0.1.7
django-tenants==2.2.3
#django-tracking-analyzer==0.2
django-user-agents==0.4.0
django-widget-tweaks==1.4.5
djangorestframework==3.10.3
easy-thumbnails==2.5
environ==1.0
feedparser==5.2.1
geoip2==2.9.0
google-api-python-client==1.7.11
google-auth==1.6.3
google-auth-httplib2==0.0.3
graphene==2.1.3
graphene-django==2.2.0
graphql-core==2.2.1
graphql-relay==0.4.5
gunicorn==19.6.0
hiredis==1.0.0
html2text==2018.1.9
httplib2==0.13.1
hyperlink==19.0.0
idna==2.8
importlib-metadata==0.20
incremental==17.5.0
Markdown==3.1.1
maxminddb==1.4.1
maxminddb-geolite2==2018.703
more-itertools==7.2.0
msgpack==0.6.1
oauth2client==4.1.3
ordereddict==1.1
packaging==19.0
phonenumbers==8.10.14
phonenumberslite==8.10.12
pilkit==2.0
Pillow==6.0.0
pluggy==0.12.0
promise==2.2.1
psycopg2==2.8.3
py==1.8.0
pyasn1==0.4.7
pyasn1-modules==0.2.6
pycparser==2.19
PyHamcrest==1.9.0
pyparsing==2.4.2
pytest==4.6.3
python-akismet==0.4.1
python-decouple==3.1
python-memcached==1.59
pytz==2018.4
PyYAML==5.1.1
requests==2.22.0
rosetta==0.3
rsa==4.0
Rx==1.6.1
singledispatch==3.4.0.3
six==1.11.0
sorl-thumbnail==12.5.0
soupsieve==1.9.3
South==1.0.2
sqlparse==0.3.0
Twisted==19.7.0
txaio==18.8.1
ua-parser==0.8.0
unicode-slugify==0.1.3
Unidecode==1.1.1
uritemplate==3.0.0
urllib3==1.25.3
user-agents==2.0
wcwidth==0.1.7
webencodings==0.5.1
whitenoise==4.1.4
zipp==0.6.0
zope.interface==4.6.0
